{Deleting equals operator in scaladoc comment breaks this comment} #SCL-6717 fixed

+++ src/org/jetbrains/plugins/scala/editor/backspaceHandler/ScalaBackspaceHandler.scala
@@ -3,14 +3,13 @@ package editor.backspaceHandler
 
 import com.intellij.codeInsight.editorActions.BackspaceHandlerDelegate
 import com.intellij.openapi.editor.Editor
-import lang.psi.api.ScalaFile
-import lang.scaladoc.lexer.docsyntax.ScaladocSyntaxElementType
-import org.jetbrains.plugins.scala.extensions
-import com.intellij.psi.{PsiElement, PsiDocumentManager, PsiFile}
 import com.intellij.psi.xml.XmlTokenType
-import lang.psi.api.expr.xml.ScXmlStartTag
-import lang.scaladoc.lexer.ScalaDocTokenType
-import lang.lexer.ScalaTokenTypes
+import com.intellij.psi.{PsiDocumentManager, PsiElement, PsiFile}
+import org.jetbrains.plugins.scala.lang.lexer.ScalaTokenTypes
+import org.jetbrains.plugins.scala.lang.psi.api.ScalaFile
+import org.jetbrains.plugins.scala.lang.psi.api.expr.xml.ScXmlStartTag
+import org.jetbrains.plugins.scala.lang.scaladoc.lexer.ScalaDocTokenType
+import org.jetbrains.plugins.scala.lang.scaladoc.lexer.docsyntax.ScaladocSyntaxElementType
 
 /**
  * User: Dmitry Naydanov
@@ -31,9 +30,12 @@ class ScalaBackspaceHandler extends BackspaceHandlerDelegate {
 
         if (element.getParent.getLastChild != element) {
           val tagToDelete = element.getParent.getLastChild
-          val textLength =
-            if (tagToDelete.getNode.getElementType != ScalaDocTokenType.DOC_BOLD_TAG) tagToDelete.getTextLength else 1
-          document.deleteString(tagToDelete.getTextOffset, tagToDelete.getTextOffset + textLength)
+
+          if (ScaladocSyntaxElementType.canClose(element.getNode.getElementType, tagToDelete.getNode.getElementType)) {
+            val textLength =
+              if (tagToDelete.getNode.getElementType != ScalaDocTokenType.DOC_BOLD_TAG) tagToDelete.getTextLength else 1
+            document.deleteString(tagToDelete.getTextOffset, tagToDelete.getTextOffset + textLength)
+          }
         } else {
           document.deleteString(element.getTextOffset, element.getTextOffset + 2)
           editor.getCaretModel.moveCaretRelatively(1, 0, false, false, false)
